refactor: extract key_names module and improve e2e test helpers
- Move is_key_name function and tests to common/key_names module
- Move WaitParams::resolve_condition to handlers::resolve_wait_condition
- Add e2e test helpers: spawn_bash_session, kill_current_session, get_session
- Add log_warn function for soft failure warnings
- Replace log_info "Note:" patterns with log_warn for clarity
- Minor code cleanup: use session_id.is_some() pattern
