+++ crates/agent-tui/src/app.rs
@@ -4,6 +4,7 @@ use clap::CommandFactory;
 use clap::Parser;
 use clap_complete::generate;
 
+use crate::common::key_names::is_key_name;
 use crate::common::{Colors, color_init};
 use crate::daemon::{DaemonError, start_daemon};
 use crate::ipc::{ClientError, DaemonClient, ensure_daemon};
@@ -21,62 +22,6 @@ mod exit_codes {
     pub const CANTCREAT: i32 = 73; // EX_CANTCREAT: can't create output
     pub const IOERR: i32 = 74; // EX_IOERR: input/output error
     pub const TEMPFAIL: i32 = 75; // EX_TEMPFAIL: temporary failure
-}
-
-/// Check if input is a recognized key name (case-insensitive).
-/// Returns true for key names like Enter, Tab, Ctrl+C, F1, etc.
-/// Returns false for text that should be typed character by character.
-fn is_key_name(input: &str) -> bool {
-    let lower = input.to_ascii_lowercase();
-
-    // Check for modifier combinations first (e.g., Ctrl+C, Alt+F4)
-    if lower.contains('+') {
-        if let Some(modifier) = lower.split('+').next() {
-            return matches!(modifier, "ctrl" | "alt" | "shift" | "meta" | "control");
-        }
-    }
-
-    // Single key names (case-insensitive)
-    matches!(
-        lower.as_str(),
-        "enter"
-            | "tab"
-            | "escape"
-            | "esc"
-            | "backspace"
-            | "delete"
-            | "arrowup"
-            | "arrowdown"
-            | "arrowleft"
-            | "arrowright"
-            | "up"
-            | "down"
-            | "left"
-            | "right"
-            | "home"
-            | "end"
-            | "pageup"
-            | "pagedown"
-            | "insert"
-            | "space"
-            | "f1"
-            | "f2"
-            | "f3"
-            | "f4"
-            | "f5"
-            | "f6"
-            | "f7"
-            | "f8"
-            | "f9"
-            | "f10"
-            | "f11"
-            | "f12"
-            | "shift"
-            | "control"
-            | "ctrl"
-            | "alt"
-            | "meta"
-    )
 }
 
 /// Application encapsulates the CLI runtime behavior.
@@ -276,8 +221,7 @@ impl Application {
                     handlers::handle_attach(ctx, attach_id.clone(), true)?
                 } else if *cleanup {
                     handlers::handle_cleanup(ctx, *all)?
-                } else if let Some(_id) = session_id {
-                    // Show details for specific session - currently just lists all
+                } else if session_id.is_some() {
                     handlers::handle_sessions(ctx)?
                 } else if *status {
                     handlers::handle_health(ctx, true)?
@@ -423,61 +367,4 @@ fn exit_code_for_client_error(error: &ClientError) -> i32 {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_is_key_name_case_insensitive() {
-        // Standard casing
-        assert!(is_key_name("Enter"));
-        assert!(is_key_name("Tab"));
-        assert!(is_key_name("Escape"));
-        assert!(is_key_name("F1"));
-
-        // Lowercase
-        assert!(is_key_name("enter"));
-        assert!(is_key_name("tab"));
-        assert!(is_key_name("escape"));
-        assert!(is_key_name("f1"));
-
-        // Uppercase
-        assert!(is_key_name("ENTER"));
-        assert!(is_key_name("TAB"));
-        assert!(is_key_name("ESCAPE"));
-        assert!(is_key_name("F1"));
-
-        // Mixed case
-        assert!(is_key_name("EnTeR"));
-        assert!(is_key_name("ArrowUp"));
-        assert!(is_key_name("arrowup"));
-        assert!(is_key_name("ARROWUP"));
-    }
-
-    #[test]
-    fn test_is_key_name_modifier_combos_case_insensitive() {
-        // Standard casing
-        assert!(is_key_name("Ctrl+c"));
-        assert!(is_key_name("Alt+F4"));
-        assert!(is_key_name("Shift+Tab"));
-
-        // Lowercase modifiers
-        assert!(is_key_name("ctrl+c"));
-        assert!(is_key_name("alt+f4"));
-        assert!(is_key_name("shift+tab"));
-
-        // Uppercase modifiers
-        assert!(is_key_name("CTRL+C"));
-        assert!(is_key_name("ALT+F4"));
-        assert!(is_key_name("SHIFT+TAB"));
-    }
-
-    #[test]
-    fn test_is_key_name_text_not_detected() {
-        // Regular text should not be detected as key names
-        assert!(!is_key_name("hello"));
-        assert!(!is_key_name("Hello World"));
-        assert!(!is_key_name("test123"));
-        assert!(!is_key_name("a")); // Single character is text, not a key
-    }
-}
+// Tests for is_key_name are in common/key_names.rs
