+++ crates/agent-tui/src/common/mod.rs
@@ -1,13 +1,14 @@
 //! Common utilities and types shared across agent-tui crates.
 //!
 //! Provides error codes, color handling, JSON extensions, string utilities,
-//! and synchronization utilities.
+//! key name validation, and synchronization utilities.
 
 #![deny(clippy::all)]
 
 mod color;
 pub mod error_codes;
 mod json_ext;
+pub mod key_names;
 mod string_utils;
 mod sync;
